Code improvement in module configuration file and update inline style
