+++ src/components/appList/AppItem.module.css
@@ -8,3 +8,63 @@
     border-radius: 5px;
     margin-top: 10px;
 }
+
+.borderInstalled {
+    border: 2px solid #21b26d;
+}
+
+.borderDisabled {
+    border: 2px solid #EAB631;
+}
+
+.borderNotInstalled {
+    border: 2px solid #F05C5C;
+}
+
+.AppItemContainerFlex {
+    display: flex;
+    align-items: center;
+}
+
+.AppItemImageStyle {
+    width: 40px;
+    height: 40px;
+}
+
+.appItemContainerMarginLeft {
+    margin-left: 30px;
+}
+
+.AppItemName {
+    font-size: 16px;
+    font-weight: bold;
+}
+
+.AppItemFlexMargin {
+    margin-top: 5px;
+    display: flex;
+    align-items: center;
+}
+
+.AppItemVersion {
+    font-style: italic;
+    color: #00000090;
+    font-size: 14px;
+}
+
+.AppItemBy {
+    margin-left: 10px;
+    font-style: italic;
+    color: #00000090;
+    font-size: 14px;
+}
+
+.AppItemStatus {
+    font-style: italic;
+    color: #00000090;
+    font-size: 14px;
+}
+
+.AppItemMarginLeft {
+    margin-left: 10px;
+}
