+++ src/components/moduleConfigurations/index.ts
@@ -1,0 +1,3 @@
+import AppListByCategory from "./AppListByCategory"
+
+export { AppListByCategory }
