+++ src/hooks/appInstallations/index.ts
@@ -2,5 +2,15 @@ import useFilterApps from './useFilterApp'
 import useGetAppListFromDHIS2 from './useGetAppListFromDHIS2'
 import useGetAppListFromDataStore from './useGetAppListFromDataStore'
 import useGetMe from './useGetMe'
-
-export { useGetMe, useGetAppListFromDHIS2, useGetAppListFromDataStore, useFilterApps }
+import useGetRightColor from './useGetRightColor'
+import useHandleFileReader from './useHandleFileReader'
+import useUpdateDataStore from './useUpdateDataStore'
+export {
+    useGetMe,
+    useFilterApps,
+    useGetRightColor,
+    useGetAppListFromDHIS2,
+    useGetAppListFromDataStore,
+    useUpdateDataStore,
+    useHandleFileReader
+}
