+++ src/hooks/moduleConfigurations/index.ts
@@ -1,0 +1,3 @@
+import useFetchDatas from "./useFetchDatas";
+
+export { useFetchDatas }
